+++ examples/daisyworld.jl
@@ -68,7 +68,7 @@ mutable struct Land <: AbstractAgent
     temperature::Float64
 end
 
-const DaisyWorld = ABM{Union{Daisy,Land}};
+const DaisyWorld = ABM{<:GridSpace,Union{Daisy,Land}};
 
 # ## World heating