+++ src/models/daisyworld.jl
@@ -16,7 +16,7 @@ mutable struct Land <: AbstractAgent
     temperature::Float64
 end
 
-const DaisyWorld = ABM{Union{Daisy, Land}};
+const DaisyWorld = ABM{<:GridSpace, Union{Daisy, Land}};
 
 """
 ``` julia